Increase pre-overlap reconstruction size to 64x64
This is done by adding the padded 32x32 reconstruction to the output of
an additional deconvolutional block. The new deconvolutional block only
uses 4 channels of the previous-layer encoding. Its role is to
approximately fit part of the object illuminated by the 'wings' of the
probe.

Probes with tails that spilled outside of the original 32x32
reconstruction box caused sketchy reconstructions. This change solves
that problem.

The new behavior can be controlled with the config parameter 'probe.big'