Add real-space MAE loss term and new 'V' dataset
The purpose of the 'V' dataset is to demo a case where supervised
training gives a better reconstruction than the PINN loss alone.

Note that both losses can be combined. However doing so does not work
well right now. This mght be because of an inconsistency between ground
truth and the model in the scaling of real-space amplitude.

Todo: real space MAE loss currently only considers amplitude. It should
also look at phase.