aider: Modified model saving to include custom objects in metadata and updated model loading to automatically retrieve and use custom objects dictionary.
# Aider chat conversation:

USER: Is there a way to save custom objects when the model is serialized but not have to manualy specify when it's re-loaded?

ASSISTANT: Yes, there is a way to save custom objects when the model is serialized and automatically load them without manually specifying each time. This can be achieved by leveraging the `custom_objects` attribute of the model's configuration. When saving the model, you can include a dictionary of custom objects in the model's metadata. Then, when loading the model, you can retrieve this dictionary and use it to automatically specify the custom objects required for loading the model. Let's implement these changes in the `ptycho/train.py` and `ptycho/inference.py` files.

### Changes to `ptycho/train.py`

We will modify the model saving part to include custom objects in the model's metadata.

```diff
--- ptycho/train.py
+++ ptycho/train.py
@@ ... @@
if save_model:
    model_path = '{}/{}'.format(out_prefix, params.get('h5_path'))
-   model.autoencoder.save(model_path, save_format="tf")
+   custom_objects = {
+       'ProbeIllumination': ProbeIllumination,
+       'IntensityScaler': IntensityScaler,
+       'IntensityScaler_inv': IntensityScaler_inv,
+       'Translation': Translation,
+       'negloglik': negloglik,
+       'realspace_loss': hh_realspace_loss
+   }
+   model.autoencoder.save(model_path, save_format="tf", custom_objects=custom_objects)
    with h5py.File(model_path, 'a') as f:
        f.attrs['intensity_scale'] = params.get('intensity_scale')
+       f.attrs['custom_objects'] = dill.dumps(custom_objects)
```

### Changes to `ptycho/inference.py`

We will modify the model loading part to automatically retrieve and use the custom objects dictionary.

```diff
--- ptycho/inference.py
+++ ptycho/inference.py
@@ ... @@
def load_pretrained_model(model_path: str) -> Model:
    """
    Load a pre-trained model from an H5 file.
    """
-   custom_objects = {
-       'ProbeIllumination': ProbeIllumination,
-       'IntensityScaler': IntensityScaler,
-       'IntensityScaler_inv': IntensityScaler_inv,
-       'Translation': Translation,
-       'negloglik': negloglik,
-       'realspace_loss': hh['realspace_loss']
-   }
+   with h5py.File(model_path, 'r') as f:
+       intensity_scale = f.attrs['intensity_scale']
+       custom_objects = dill.loads(f.attrs['custom_objects'])
    model = tf_load_model(model_path, custom_objects=custom_objects)
-   with h5py.File(model_path, 'r') as f:
-       intensity_scale = f.attrs['intensity_scale']
    params.set('intensity_scale', intensity_scale)
    return model
```

These changes will streamline the process of saving and loading models with custom objects, making it more efficient and less error-prone.